+++ resources/install_apt.sh
@@ -12,6 +12,8 @@ installVer='14' 	#NodeJS major version to be installed
 pre
 step 0 "Vérification des droits"
 silent sudo killall easy-server
+silent sudo killall bosch-xmpp
+
 DIRECTORY="/var/www"
 if [ ! -d "$DIRECTORY" ]; then
 	silent sudo mkdir $DIRECTORY
@@ -25,26 +27,19 @@ try sudo DEBIAN_FRONTEND=noninteractive apt-get install -y lsb-release
 #install nodejs, steps 10->50
 . ${BASEDIR}/install_nodejs.sh ${installVer}
 
-step 60 "Nettoyage anciens modules"
-sudo npm ls -g --depth 0 2>/dev/null | grep "nefit-easy-http-server@" >/dev/null 
-if [ $? -ne 1 ]; then
-  echo "[Suppression easy global"
-  silent sudo npm rm -g nefit-easy-http-server
-fi
 cd ${BASEDIR};
 #remove old local modules
 sudo rm -rf node_modules &>/dev/null
 sudo rm -f package-lock.json &>/dev/null
 
-step 70 "Installation de Nefit Easy Server, veuillez patienter svp"
+step 70 "Installation de Bosch XMPP, veuillez patienter svp"
 #need to be sudoed because of recompil
 silent sudo mkdir node_modules
 silent sudo chown -R www-data:www-data .
 
-sudo npm install -g nefit-easy-http-server
-serverversion=`easy-server -v`;
-step 80 "Nefit Easy HTTP Server version ${serverversion} installé."
-
+sudo npm install -g bosch-xmpp
+serverversion=`bosch-xmpp -v`;
+step 80 "Bosch XMPP version ${serverversion} installé."
 
 silent sudo chown -R www-data:www-data .